+++ app/src/components/manage/ManageCard.jsx
@@ -1,0 +1,6 @@
+import Button from "../ui/Button";
+
+export default function ManageCard(){
+
+}
+
