+++ app/src/pages/Manage.jsx
@@ -1,7 +1,8 @@
-
+import ManagePanel from "../components/manage/ManagePanel";
 export default function Manage(){
     return(
         <>
+        <ManagePanel />
         </>
     );
 }
